feat: v2 start on enums
