+++ src/postgres/models/mod.rs
@@ -1,0 +1,1 @@
+pub mod postgres_enum;
